move AuctioDatePickerValue to types, simplify AuctionDatePicker handleSelect
