+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/NftForm.types.ts
@@ -1,5 +1,3 @@
-import { AuctionDatePickerValue } from '@/components/_inputs/AuctionDatePicker'
-
 export enum AuctionDuration {
   NoExpiration = 'No expiration date',
   OneDay = '1 day',
@@ -9,6 +7,17 @@ export enum AuctionDuration {
 }
 
 export type Listing = 'Auction' | 'Fixed price' | 'Not for sale' | undefined
+
+export type AuctionDatePickerValue =
+  | {
+      type: 'date'
+      date: Date
+    }
+  | {
+      type: 'duration'
+      durationDays: number | null
+    }
+  | null
 
 export type AuctionDate = {
   startDate: AuctionDatePickerValue
